refactor: extract search params validation into hook
