+++ src/domains/transaction/hooks/use-transaction-url.ts
@@ -1,13 +1,6 @@
 import { Contracts } from "@ardenthq/sdk-profiles";
 import { generatePath } from "react-router";
-import { useTranslation } from "react-i18next";
 import { ProfilePaths } from "@/router/paths";
-
-interface RequiredOptions {
-	network?: string;
-	coin?: string;
-	nethash?: string;
-}
 
 const generateSendTransferPath = (profile: Contracts.IProfile, url: string) => {
 	const path = generatePath(ProfilePaths.SendTransfer, { profileId: profile.id() });
@@ -18,49 +11,7 @@ const generateSendTransferPath = (profile: Contracts.IProfile, url: string) => {
 
 const urlSearchParameters = (url: string) => new URL(url.replace("#", "/")).searchParams;
 
-export const useTransactionURL = () => {
-	const { t } = useTranslation();
-
-	const validateTransferURLParameters = (url: string, requiredOptions: RequiredOptions) => {
-		let searchParameters: URLSearchParams;
-
-		try {
-			searchParameters = urlSearchParameters(url);
-		} catch {
-			return t("TRANSACTION.INVALID_URL");
-		}
-
-		if (!searchParameters.get("coin")) {
-			return t("TRANSACTION.VALIDATION.COIN_MISSING");
-		}
-
-		const network = searchParameters.get("network");
-		const nethash = searchParameters.get("nethash");
-
-		if (network && !["ark.devnet", "ark.mainnet"].includes(network)) {
-			return t("TRANSACTION.VALIDATION.NETWORK_INVALID");
-		}
-
-		if (!network && !nethash) {
-			return t("TRANSACTION.VALIDATION.NETWORK_OR_NETHASH_MISSING");
-		}
-
-		if (!!network && network !== requiredOptions.network) {
-			return t("TRANSACTION.VALIDATION.NETWORK_MISMATCH");
-		}
-
-		if (!!nethash && nethash !== requiredOptions.nethash) {
-			return t("TRANSACTION.VALIDATION.NETWORK_MISMATCH");
-		}
-
-		if (searchParameters.get("coin") !== requiredOptions.coin) {
-			return t("TRANSACTION.VALIDATION.COIN_MISMATCH");
-		}
-	};
-
-	return {
-		generateSendTransferPath,
-		urlSearchParameters,
-		validateTransferURLParameters,
-	};
-};
+export const useTransactionURL = () => ({
+	generateSendTransferPath,
+	urlSearchParameters,
+});
